# shared-contracts/target-operational-v2.openapi.json
{
  "components": {
    "schemas": {
...
      "CursorPageInfo": {
        "additionalProperties": false,
        "properties": {
          "hasMore": {
            "title": "Hasmore",
            "type": "boolean"
          },
          "limit": {
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "nextCursor": {
            "anyOf": [
              {
                "maxLength": 1024,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nextcursor"
          },
          "returnedCount": {
            "minimum": 0.0,
            "title": "Returnedcount",
            "type": "integer"
          }
        },
        "required": [
          "limit",
          "returnedCount",
          "hasMore",
          "nextCursor"
        ],
        "title": "CursorPageInfo",
        "type": "object"
      },
...
      "FinalReportArtifactDetail": {
        "additionalProperties": false,
        "properties": {
          "artifactId": {
            "format": "uuid",
            "title": "Artifactid",
            "type": "string"
          },
          "contentHash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contenthash"
          },
          "contractVersion": {
            "const": 2,
            "default": 2,
            "title": "Contractversion",
            "type": "integer"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "downloadAvailable": {
            "title": "Downloadavailable",
            "type": "boolean"
          },
          "finalReportVersionId": {
            "format": "uuid",
            "title": "Finalreportversionid",
            "type": "string"
          },
          "generatedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generatedat"
          },
          "generatedByAccountId": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generatedbyaccountid"
          },
          "generationAttempts": {
            "minimum": 0.0,
            "title": "Generationattempts",
            "type": "integer"
          },
          "lastErrorCode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lasterrorcode"
          },
          "mimeType": {
            "title": "Mimetype",
            "type": "string"
          },
          "reportFinalizationId": {
            "format": "uuid",
            "title": "Reportfinalizationid",
            "type": "string"
          },
          "sizeBytes": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sizebytes"
          },
          "status": {
            "enum": [
              "pending",
              "ready",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          },
          "templateVersion": {
            "title": "Templateversion",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          }
        },
        "required": [
          "reportFinalizationId",
          "finalReportVersionId",
          "artifactId",
          "status",
          "templateVersion",
          "mimeType",
          "contentHash",
          "generationAttempts",
          "lastErrorCode",
          "generatedAt",
          "generatedByAccountId",
          "downloadAvailable",
          "createdAt",
          "updatedAt"
        ],
        "title": "FinalReportArtifactDetail",
        "type": "object"
      },
...
      "FinalReportPage": {
        "additionalProperties": false,
        "properties": {
          "classification": {
            "const": "official",
            "default": "official",
            "title": "Classification",
            "type": "string"
          },
          "contractVersion": {
            "const": 2,
            "default": 2,
            "title": "Contractversion",
            "type": "integer"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/FinalReportListItem"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "$ref": "#/components/schemas/CursorPageInfo"
          }
        },
        "required": [
          "items",
          "page"
        ],
        "title": "FinalReportPage",
        "type": "object"
      },
...
      "FinalizationAcknowledgement": {
        "additionalProperties": false,
        "properties": {
          "acknowledgedAt": {
            "format": "date-time",
            "title": "Acknowledgedat",
            "type": "string"
          },
          "commandId": {
            "format": "uuid",
            "title": "Commandid",
            "type": "string"
          },
          "contractVersion": {
            "const": 2,
            "title": "Contractversion",
            "type": "integer"
          },
          "disposition": {
            "enum": [
              "created",
              "replayed"
            ],
            "title": "Disposition",
            "type": "string"
          },
          "payloadHash": {
            "title": "Payloadhash",
            "type": "string"
          },
          "resource": {
            "$ref": "#/components/schemas/FinalizationAcknowledgementResource"
          }
        },
        "required": [
          "contractVersion",
          "commandId",
          "disposition",
          "payloadHash",
          "acknowledgedAt",
          "resource"
        ],
        "title": "FinalizationAcknowledgement",
        "type": "object"
      },
...
      "FinalizeReportsCommand": {
        "additionalProperties": false,
        "properties": {
          "commandId": {
            "format": "uuid",
            "title": "Commandid",
            "type": "string"
          },
          "contractVersion": {
            "const": 2,
            "title": "Contractversion",
            "type": "integer"
          },
          "expectedVersion": {
            "minimum": 0.0,
            "title": "Expectedversion",
            "type": "integer"
          },
          "idempotencyKey": {
            "maxLength": 240,
            "minLength": 8,
            "pattern": "^final-report:[A-Za-z0-9._-]+:[A-Za-z0-9._-]+$",
            "title": "Idempotencykey",
            "type": "string"
          },
          "occurredAt": {
            "format": "date-time",
            "title": "Occurredat",
            "type": "string"
          },
          "payload": {
            "$ref": "#/components/schemas/FinalizeReportsPayload"
          }
        },
        "required": [
          "contractVersion",
          "commandId",
          "idempotencyKey",
          "occurredAt",
          "expectedVersion",
          "payload"
        ],
        "title": "FinalizeReportsCommand",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OperationalAlertPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/OperationalAlertSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "$ref": "#/components/schemas/CursorPageInfo"
          }
        },
        "required": [
          "items",
          "page"
        ],
        "title": "OperationalAlertPage",
        "type": "object"
      },
      "OperationalAlertStatusUpdate": {
        "properties": {
          "status": {
            "enum": [
              "New",
              "In Review",
              "Resolved"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "OperationalAlertStatusUpdate",
        "type": "object"
      },
      "OperationalAlertSummary": {
        "properties": {
          "enterprise": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enterprise"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "owner": {
            "enum": [
              "IT",
              "Admin",
              "System"
            ],
            "title": "Owner",
            "type": "string"
          },
          "requester": {
            "title": "Requester",
            "type": "string"
          },
          "requiredAction": {
            "title": "Requiredaction",
            "type": "string"
          },
          "resolutionMode": {
            "enum": [
              "On-site Visit Required",
              "In-system Action",
              "Staff Follow-up",
              "Remote Review",
              "Admin Monitoring",
              "Automatic Health Recovery"
            ],
            "title": "Resolutionmode",
            "type": "string"
          },
          "severity": {
            "enum": [
              "Info",
              "Warning",
              "Critical"
            ],
            "title": "Severity",
            "type": "string"
          },
          "status": {
            "enum": [
              "New",
              "In Review",
              "Resolved"
            ],
            "title": "Status",
            "type": "string"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "time": {
            "title": "Time",
            "type": "string"
          },
          "type": {
            "enum": [
              "Maintenance Request",
              "Password Reset Request",
              "Submission Delay",
              "Threshold Breach",
              "Foot Traffic Alert",
              "Occupancy Spike",
              "Failed Login Threshold",
              "Sync Delay"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "type",
          "severity",
          "requester",
          "summary",
          "requiredAction",
          "resolutionMode",
          "status",
          "owner",
          "time"
        ],
        "title": "OperationalAlertSummary",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "TANAW API",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/operational/alerts": {
      "get": {
        "operationId": "list_alerts_operational_alerts_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 1024,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationalAlertPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Alerts",
        "tags": [
          "alerts"
        ]
      }
    },
    "/operational/alerts/{alert_code}/status": {
      "patch": {
        "operationId": "update_alert_status_operational_alerts__alert_code__status_patch",
        "parameters": [
          {
            "in": "path",
            "name": "alert_code",
            "required": true,
            "schema": {
              "title": "Alert Code",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OperationalAlertStatusUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationalAlertSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Alert Status",
        "tags": [
          "alerts"
        ]
      }
    },
...
    "/operational/reports/finalizations/v2": {
      "get": {
        "operationId": "list_final_reports_v2_operational_reports_finalizations_v2_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 1024,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "reportingPeriodId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Reportingperiodid"
            }
          },
          {
            "in": "query",
            "name": "scopeType",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "citywide",
                    "barangay",
                    "enterprise_selection"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scopetype"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FinalReportPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Final Reports V2",
        "tags": [
          "final-reports-v2"
        ]
      },
      "post": {
        "operationId": "finalize_reports_v2_operational_reports_finalizations_v2_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FinalizeReportsCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FinalizationAcknowledgement"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Finalize Reports V2",
        "tags": [
          "final-reports-v2"
        ]
      }
    },
    "/operational/reports/finalizations/{report_finalization_id}/artifacts/{artifact_id}/download/v2": {
      "get": {
        "operationId": "download_final_report_artifact_v2_operational_reports_finalizations__report_finalization_id__artifacts__artifact_id__download_v2_get",
        "parameters": [
          {
            "in": "path",
            "name": "report_finalization_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Report Finalization Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "artifact_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Artifact Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Download Final Report Artifact V2",
        "tags": [
          "final-reports-v2"
        ]
      }
    },
    "/operational/reports/finalizations/{report_finalization_id}/artifacts/{artifact_id}/v2": {
      "get": {
        "operationId": "read_final_report_artifact_v2_operational_reports_finalizations__report_finalization_id__artifacts__artifact_id__v2_get",
        "parameters": [
          {
            "in": "path",
            "name": "report_finalization_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Report Finalization Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "artifact_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Artifact Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FinalReportArtifactDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read Final Report Artifact V2",
        "tags": [
          "final-reports-v2"
        ]
      }
    },
    "/operational/reports/finalizations/{report_finalization_id}/v2": {
      "get": {
        "operationId": "read_final_report_v2_operational_reports_finalizations__report_finalization_id__v2_get",
        "parameters": [
          {
            "in": "path",
            "name": "report_finalization_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Report Finalization Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "versionId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Versionid"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FinalReportDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read Final Report V2",
        "tags": [
          "final-reports-v2"
        ]
      }
    },
...
  }
}